# docs/openapi/quotes-v0.openapi.yaml
openapi: 3.1.1
info:
  title: Quotes.Api | v0
  description: "Quote catalog for the Aspire Quotes platform.\n\n**Two transports,\
    \ one contract**: `v0` (MVC controllers) and `v1` (minimal APIs)\npublish the\
    \ same operations, payloads and error envelope; parity is enforced by\ntests.\
    \ `v0` exists to demonstrate that transport style is a swappable detail of the\n\
    architecture — new integrations should prefer `v1`.\n\nTypical use:\n\n1. Obtain\
    \ a bearer JWT from the Auth API: `POST /api/v1/auth/login` (development\n   users:\
    \ `jrb`/`supersecret` with both scopes, `reader`/`readsecret` read-only).\n2.\
    \ Send `Authorization: Bearer {accessToken}` to every operation below — reads\n\
    \   require the `quotes:read` scope claim, create requires `quotes:write`; a valid\n\
    \   token without the scope answers 403.\n3. `POST /quotes` first: the catalog\
    \ starts empty. Then browse with `GET /quotes`,\n   `GET /quotes/{id}` or `GET\
    \ /quotes/random`.\n\nCross-cutting behavior:\n\n- Every error response is RFC\
    \ 9457 `application/problem+json` with `errorCode` and\n  `correlationId` extensions.\n\
    - Pagination is 1-based (`page` from 1, `pageSize` between 1 and 100, default\
    \ 20).\n- Send `X-Correlation-Id` to correlate calls; it is echoed on every response."
  version: 1.0.0
servers:
- url: /
paths:
  /api/v0/quotes/random:
    get:
      tags:
      - Quotes
      summary: Returns a random quote from the catalog.
      description: 'Requires a bearer JWT issued by the Auth API with the `quotes:read`
        scope

        (`POST /api/v1/auth/login`); a valid token without the scope answers 403.
        The

        catalog starts empty, so calls before the first create answer 404. Send

        `X-Correlation-Id` to correlate calls; it is echoed on every response and

        embedded in problem details.'
      responses:
        '401':
          description: Missing or invalid bearer token (errorCode `auth.token_missing`
            or `auth.token_invalid`).
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: https://tools.ietf.org/html/rfc9110#section-15.5.2
                title: Unauthorized
                status: 401
                detail: A valid bearer token is required.
                errorCode: auth.token_invalid
                correlationId: 5c1f4a0e9d2b7386a4c0b1e8d3f69a27
        '403':
          description: The token lacks the `quotes:read` scope.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: https://tools.ietf.org/html/rfc9110#section-15.5.4
                title: Forbidden
                status: 403
                detail: The access token is missing the required scope (quotes:read
                  or quotes:write).
        '200':
          description: A random quote.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuoteResponseDto'
        '404':
          description: The catalog is empty (errorCode `quote.not_found`).
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: https://tools.ietf.org/html/rfc9110#section-15.5.5
                title: Not Found
                status: 404
                detail: Quote not found.
                errorCode: quote.not_found
                correlationId: 5c1f4a0e9d2b7386a4c0b1e8d3f69a27
      security:
      - Bearer: []
  /api/v0/quotes:
    get:
      tags:
      - Quotes
      summary: Lists one page of the quote catalog in stable order.
      description: 'Requires a bearer JWT issued by the Auth API with the `quotes:read`
        scope

        (`POST /api/v1/auth/login`); a valid token without the scope answers 403.

        Pagination is 1-based: `page` starts at 1 and `pageSize` is capped at 100

        (default 20); values outside those ranges answer 400 with

        `quote.invalid_page_request`. Pages beyond the last return an empty

        `items` array, not an error. Send `X-Correlation-Id` to correlate calls;

        it is echoed on every response and embedded in problem details.'
      operationId: ListQuotesV0
      parameters:
      - name: page
        in: query
        description: 1-based page number (minimum 1).
        schema:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
          - integer
          - string
          format: int32
          default: 1
        example: 2
      - name: pageSize
        in: query
        description: Items per page, between 1 and 100 (default 20).
        schema:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
          - integer
          - string
          format: int32
          default: 20
        example: 20
      responses:
        '401':
          description: Missing or invalid bearer token (errorCode `auth.token_missing`
            or `auth.token_invalid`).
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: https://tools.ietf.org/html/rfc9110#section-15.5.2
                title: Unauthorized
                status: 401
                detail: A valid bearer token is required.
                errorCode: auth.token_invalid
                correlationId: 5c1f4a0e9d2b7386a4c0b1e8d3f69a27
        '403':
          description: The token lacks the `quotes:read` scope.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: https://tools.ietf.org/html/rfc9110#section-15.5.4
                title: Forbidden
                status: 403
                detail: The access token is missing the required scope (quotes:read
                  or quotes:write).
        '200':
          description: The requested page as `{ items, page, pageSize, totalItems,
            totalPages }`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuotePageResponseDto'
        '400':
          description: Page or pageSize outside the allowed range (errorCode `quote.invalid_page_request`).
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/HttpValidationProblemDetails'
              example:
                type: https://tools.ietf.org/html/rfc9110#section-15.5.1
                title: One or more validation errors occurred.
                status: 400
                errorCode: quote.invalid_page_request
                correlationId: 5c1f4a0e9d2b7386a4c0b1e8d3f69a27
                errors:
                  quote.invalid_page_request:
                  - The requested page or page size is outside the allowed range.
      security:
      - Bearer: []
    post:
      tags:
      - Quotes
      summary: Adds a quote to the catalog.
      description: 'Requires a bearer JWT issued by the Auth API with the `quotes:write`
        scope

        (`POST /api/v1/auth/login`); read-only tokens answer 403. Catalog rules beyond

        the schema limits: text between 12 and 280 characters with at least 3 words,
        ending

        with ''.'', ''!'' or ''?''; author between 2 and 80 characters (letters, spaces,
        hyphens,

        apostrophes and periods) and different from the text. Near-duplicates (same

        fingerprint, for example only punctuation changed) answer 409. The 201 response

        carries the created quote and a `Location` header addressing it. Send

        `X-Correlation-Id` to correlate calls; it is echoed on every response and

        embedded in problem details.'
      requestBody:
        description: The quote text and its author.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateQuoteRequestDto'
          text/json:
            schema:
              $ref: '#/components/schemas/CreateQuoteRequestDto'
          application/*+json:
            schema:
              $ref: '#/components/schemas/CreateQuoteRequestDto'
        required: true
      responses:
        '401':
          description: Missing or invalid bearer token (errorCode `auth.token_missing`
            or `auth.token_invalid`).
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: https://tools.ietf.org/html/rfc9110#section-15.5.2
                title: Unauthorized
                status: 401
                detail: A valid bearer token is required.
                errorCode: auth.token_invalid
                correlationId: 5c1f4a0e9d2b7386a4c0b1e8d3f69a27
        '403':
          description: The token lacks the `quotes:write` scope.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: https://tools.ietf.org/html/rfc9110#section-15.5.4
                title: Forbidden
                status: 403
                detail: The access token is missing the required scope (quotes:read
                  or quotes:write).
        '201':
          description: Quote created; the `Location` header addresses the new quote.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuoteResponseDto'
        '400':
          description: Catalog rule violation (errorCode `quote.text_too_short`, `quote.text_too_long`,
            `quote.text_needs_more_words`, `quote.text_must_end_with_punctuation`,
            `quote.author_too_short`, `quote.author_too_long`, `quote.author_invalid_characters`
            or `quote.author_equals_text`).
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/HttpValidationProblemDetails'
              example:
                type: https://tools.ietf.org/html/rfc9110#section-15.5.1
                title: One or more validation errors occurred.
                status: 400
                errorCode: quote.text_too_short
                correlationId: 5c1f4a0e9d2b7386a4c0b1e8d3f69a27
                errors:
                  quote.text_too_short:
                  - Quote text must be at least 12 characters.
        '409':
          description: A quote with the same meaning already exists (errorCode `quote.duplicate_fingerprint`).
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: https://tools.ietf.org/html/rfc9110#section-15.5.10
                title: Conflict
                status: 409
                detail: A quote with the same meaning already exists.
                errorCode: quote.duplicate_fingerprint
                correlationId: 5c1f4a0e9d2b7386a4c0b1e8d3f69a27
      security:
      - Bearer: []
  /api/v0/quotes/{id}:
    get:
      tags:
      - Quotes
      summary: Returns one quote by id.
      description: 'Requires a bearer JWT issued by the Auth API with the `quotes:read`
        scope

        (`POST /api/v1/auth/login`); a valid token without the scope answers 403.
        Ids

        come from create responses and list items. Send `X-Correlation-Id` to

        correlate calls; it is echoed on every response and embedded in problem details.'
      operationId: GetQuoteByIdV0
      parameters:
      - name: id
        in: path
        description: Quote identifier as returned by create or list.
        required: true
        schema:
          type: string
        example: 3f2b8a9c1d4e5f6a7b8c9d0e1f2a3b4c
      responses:
        '401':
          description: Missing or invalid bearer token (errorCode `auth.token_missing`
            or `auth.token_invalid`).
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: https://tools.ietf.org/html/rfc9110#section-15.5.2
                title: Unauthorized
                status: 401
                detail: A valid bearer token is required.
                errorCode: auth.token_invalid
                correlationId: 5c1f4a0e9d2b7386a4c0b1e8d3f69a27
        '403':
          description: The token lacks the `quotes:read` scope.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: https://tools.ietf.org/html/rfc9110#section-15.5.4
                title: Forbidden
                status: 403
                detail: The access token is missing the required scope (quotes:read
                  or quotes:write).
        '200':
          description: The requested quote.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuoteResponseDto'
        '404':
          description: No quote matches the id (errorCode `quote.not_found`).
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: https://tools.ietf.org/html/rfc9110#section-15.5.5
                title: Not Found
                status: 404
                detail: Quote not found.
                errorCode: quote.not_found
                correlationId: 5c1f4a0e9d2b7386a4c0b1e8d3f69a27
      security:
      - Bearer: []
components:
  schemas:
    CreateQuoteRequestDto:
      required:
      - text
      - author
      type: object
      properties:
        text:
          maxLength: 280
          type: string
          description: Quote body text.
        author:
          maxLength: 80
          type: string
          description: Attributed author of the quote.
      example:
        text: Talk is cheap. Show me the code.
        author: Linus Torvalds
    HttpValidationProblemDetails:
      type: object
      properties:
        type:
          type:
          - 'null'
          - string
        title:
          type:
          - 'null'
          - string
        status:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
          - 'null'
          - integer
          - string
          format: int32
        detail:
          type:
          - 'null'
          - string
        instance:
          type:
          - 'null'
          - string
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
    ProblemDetails:
      type: object
      properties:
        type:
          type:
          - 'null'
          - string
        title:
          type:
          - 'null'
          - string
        status:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
          - 'null'
          - integer
          - string
          format: int32
        detail:
          type:
          - 'null'
          - string
        instance:
          type:
          - 'null'
          - string
    QuotePageResponseDto:
      required:
      - items
      - page
      - pageSize
      - totalItems
      - totalPages
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/QuoteResponseDto'
          description: Quotes on this page, in stable catalog order.
        page:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
          - integer
          - string
          description: 1-based page number this response represents.
          format: int32
        pageSize:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
          - integer
          - string
          description: Number of items per page that was requested.
          format: int32
        totalItems:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
          - integer
          - string
          description: Total number of quotes in the catalog.
          format: int32
        totalPages:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
          - integer
          - string
          description: Total number of pages at the requested page size.
          format: int32
      example:
        items:
        - id: 3f2b8a9c1d4e5f6a7b8c9d0e1f2a3b4c
          text: Talk is cheap. Show me the code.
          author: Linus Torvalds
        page: 1
        pageSize: 20
        totalItems: 1
        totalPages: 1
    QuoteResponseDto:
      required:
      - id
      - text
      - author
      type: object
      properties:
        id:
          type: string
          description: Stable quote identifier.
        text:
          type: string
          description: Quote body text.
        author:
          type: string
          description: Attributed author of the quote.
      example:
        id: 3f2b8a9c1d4e5f6a7b8c9d0e1f2a3b4c
        text: Talk is cheap. Show me the code.
        author: Linus Torvalds
  securitySchemes:
    Bearer:
      type: http
      description: Enter the bearer token issued by the Auth API (POST /api/v1/auth/login).
      scheme: bearer
      bearerFormat: JWT
...